+++ src/table/api/zh-cn.js
@@ -31,6 +31,18 @@ module.exports = [
                 default: 'false'
             },
             {
+                name: 'thyHeaderFixed',
+                description: '是否表头固定，若设置为 true， 需要同步设置 thyHeight',
+                type: 'boolean',
+                default: 'false'
+            },
+            {
+                name: 'thyHeight',
+                description: '表格高度',
+                type: `number | string`,
+                default: ''
+            },
+            {
                 name: 'thySize',
                 description: `表格大小，'xs' | 'sm' | 'md' | 'lg' | 'xlg' 默认为 'md'`,
                 type: `'xs' | 'sm' | 'md' | 'lg' | 'xlg'`,